This fixes #7533.  The SQL server tables have been upgraded to doubles instead of floats.

+++ src/php/db/History.php
@@ -219,7 +219,7 @@ abstract class History extends Table
         for ($i = 0; $i < $this->datacols; $i++) {
             $this->sqlColumns["Data$i"] = array(
                 "Name" => "Data$i",
-                "Type" => "float",
+                "Type" => "double",
                 "Default" => null,
                 "Null" => true,
             );
